cambios Exceptions

+++ cuenta-movimiento-service/src/main/java/com/pruebatecnica/cuentamovimientoservice/Controller/CuentaController.java
@@ -2,12 +2,14 @@ package com.pruebatecnica.cuentamovimientoservice.Controller;
 
 
 import com.pruebatecnica.cuentamovimientoservice.Entity.Cuenta;
+import com.pruebatecnica.cuentamovimientoservice.Exceptions.ClienteNotFoundException;
 import com.pruebatecnica.cuentamovimientoservice.Repository.CuentaRepository;
 import com.pruebatecnica.cuentamovimientoservice.Service.MessageProducer;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.support.SessionStatus;
 
 import javax.swing.text.html.parser.Entity;
 import java.util.List;
@@ -16,14 +18,12 @@ import java.util.List;
 @RequestMapping("/api/cuentas")
 public class CuentaController {
 
-    //private final MessageConsumer messageConsumer;
     private CuentaRepository cuentaRepository;
     private  MessageProducer messageProducer;
 
     public CuentaController(CuentaRepository cuentaRepository, MessageProducer messageProducer){
         this.cuentaRepository = cuentaRepository;
         this.messageProducer = messageProducer;
-        //this.messageConsumer = messageConsumer;
     }
 
     @GetMapping
@@ -38,27 +38,24 @@ public class CuentaController {
 
     @GetMapping("/{id}")
     public Cuenta getCuentaById(@PathVariable Long id){
-        return  cuentaRepository.findById(id).orElseThrow(()->new RuntimeException("Cuenta no encontrada"));
+        return  cuentaRepository.findById(id).orElseThrow(()->new ClienteNotFoundException("Cuenta no encontrada"));
     }
 
     @DeleteMapping("/{id}")
     public void deleteCuenta(@PathVariable Long id){
+        if(!cuentaRepository.existsById(id)){
+            throw  new ClienteNotFoundException("Cliente no encontrdo con ID:"+id);
+        }
           cuentaRepository.deleteById(id);
     }
 
     @PostMapping("/send")
     public void sendMessage(@RequestBody String message) {
-        Cuenta cuenta = new Cuenta();
-        cuenta.setId(1L);
-        cuenta.setTipoCuenta("Ahorro");
-        cuenta.setNumeeroCuenta("1234456643");
-        messageProducer.sendMessage(cuenta.toString());
+        //Cuenta cuenta = new Cuenta();
+        //cuenta.setId(1L);
+        //cuenta.setTipoCuenta("Ahorro");
+        //cuenta.setNumeeroCuenta("1234456643");
+        messageProducer.sendMessage(message);
+        //messageProducer.sendMessage(cuenta.toString());
     }
-/*
-    @GetMapping("/read")
-    public String readMessage() {
-        return messageConsumer.receiveMessage("message");
-    }
-
- */
 }
